+++ ppdet/modeling/__init__.py
@@ -1,11 +1,13 @@
 from . import ops
 from . import anchor
+from . import mask
 from . import backbone
 from . import head
 from . import architecture
 
 from .ops import *
 from .anchor import *
+from .mask import *
 from .backbone import *
 from .head import *
 from .architecture import *
